Refactor basic deep link recipe
Refactored to use UriDeepLinkMatcher.

Partially addresses #301
