+++ app/src/main/java/com/example/nav3recipes/deeplink/handlerequests/staticuri/MainActivity.kt
@@ -10,8 +10,8 @@ import androidx.navigation3.runtime.rememberNavBackStack
 import androidx.navigation3.ui.NavDisplay
 import com.example.nav3recipes.common.deeplink.EntryScreen
 import com.example.nav3recipes.common.deeplink.TextContent
-import com.example.nav3recipes.deeplink.handlerequests.basic.HomeKey
-import com.example.nav3recipes.deeplink.handlerequests.basic.NavRecipeKey
+import com.example.nav3recipes.deeplink.handlerequests.uriarguments.HomeKey
+import com.example.nav3recipes.deeplink.handlerequests.uriarguments.NavRecipeKey
 import androidx.navigation3.runtime.deeplink.DeepLinkRequest
 import androidx.navigation3.runtime.deeplink.DeepLinkUri
 import androidx.navigation3.runtime.deeplink.UriDeepLinkMatcher
