+++ app/src/main/java/com/example/nav3recipes/deeplink/handlerequests/staticuri/StaticUriDeepLinkActivity.kt
@@ -9,7 +9,7 @@ import androidx.lifecycle.compose.dropUnlessResumed
 import com.example.nav3recipes.common.deeplink.EntryScreen
 import com.example.nav3recipes.common.deeplink.PaddedButton
 import com.example.nav3recipes.common.deeplink.TextContent
-import com.example.nav3recipes.deeplink.handlerequests.basic.ui.PATH_BASE
+import com.example.nav3recipes.deeplink.handlerequests.uriarguments.ui.PATH_BASE
 import com.example.nav3recipes.ui.setEdgeToEdgeConfig
 
 const val HOME_URI = "$PATH_BASE/home"
